[indeterminate-state-checkbox] Move `Checked` to central location in types.ts
